Factor out a `LockDownloader`.
This both makes the code easier to read (arguable) and allows re-use of
this logic for downloading locked build system artifacts.
